Rename components:
start-servant -> commence-core
prototype-hs-lib -> commence-interactive-state

@thusc since we'd like to namespace these packages under commence, for
me the names of start-servant vs. prototype-hs-lib were getting too
confusing.

To aid in organisation, I've therefore taken the liberty to further
divide commence into 2 parts:

1. core
This contains the core of commence: things like error handling,
abstractions around storage etc.

2. interactive-state
An additional feature, that aids in desiging repls. This may not be
part of the core since we may not need to implement this in each
project.

--

I hope this is OK with you.